requirements.txt
rich
glob2
tqdm
torch
torchaudio
torchvision
numpy
scipy
pyyaml
transformers
